Added ability to calculate arena position
Based off of maths outlined here: http://i.imgur.com/d1d8f29.png
Calculates the position of the robot in the arena by using arena markers
as known points. Coordinates are in metres from the top left of the
arena (Northwest of the diagram on pg. 14 of the rulebook:
https://www.studentrobotics.org/resources/2013/rulebook.pdf)